ADD: QueryCreator构建sql语句元素的util类
