+++ protoframework-data/src/main/java/org/protoframework/sql/QueryCreator.java
@@ -1,0 +1,55 @@
+package org.protoframework.sql;
+
+import org.protoframework.sql.clause.*;
+import org.protoframework.sql.expr.TableColumn;
+import org.protoframework.sql.expr.Value;
+
+/**
+ * @author: yuanwq
+ * @date: 2018/7/16
+ */
+public abstract class QueryCreator {
+
+  public static ISqlValue sqlValue(Object value) {
+    if (value instanceof ISqlValue) {
+      return (ISqlValue) value;
+    }
+    return new Value(value);
+  }
+
+  public static TableColumn column(String field) {
+    return new TableColumn(field);
+  }
+
+  public static SelectClause select() {
+    return new SelectClause();
+  }
+
+  public static FromClause from(String tableName) {
+    return new FromClause(table(tableName));
+  }
+
+  public static ITableRef table(String tableName) {
+    return new SimpleTable(tableName);
+  }
+
+  public static WhereClause where() {
+    return new WhereClause();
+  }
+
+  public static GroupByClause groupBy() {
+    return new GroupByClause();
+  }
+
+  public static OrderByClause orderBy() {
+    return new OrderByClause();
+  }
+
+  public static PaginationClause.Builder pagination(int limit) {
+    return PaginationClause.newBuilder(limit);
+  }
+
+  public static SetClause set() {
+    return new SetClause();
+  }
+}
